+++ packages/components/molecules/f-skeleton-loader/src/components/skeletons/index.js
@@ -1,3 +1,8 @@
 export { default as TextBlock } from './Text.vue';
 export { default as Heading } from './Heading.vue';
 export { default as Sentence } from './Sentence.vue';
+export { default as ImageBlock } from './Image.vue';
+export { default as Rating } from './Rating.vue';
+export { default as CarouselCardRestaurantHeader } from './CarouselCardRestaurantHeader.vue';
+export { default as CarouselCardRestaurantData } from './CarouselCardRestaurantData.vue';
+export { default as RestaurantCarouselCard } from './RestaurantCarouselCard.vue';