+++ packages/components/organisms/f-checkout/test-utils/component-objects/f-checkout.component.js
@@ -10,7 +10,14 @@ const {
     KNOB_BUTTON,
     SWITCH_USER_LINK,
     GUEST_CHECKOUT_HEADER,
-    GUEST_CHECKOUT_LOGIN_BUTTON
+    GUEST_CHECKOUT_LOGIN_BUTTON,
+    PRE_ORDER_WARNING,
+    CHECKOUT_ERROR_MESSAGE,
+    RETRY_BUTTON,
+    ERROR_PAGE_COMPONENT,
+    ERROR_PAGE_HEADING,
+    ERROR_PAGE_DESCRIPTION,
+    ERROR_PAGE_IMAGE
 } = require('./f-checkout-selectors');
 
 module.exports = class Checkout extends Page {
@@ -33,6 +40,20 @@ module.exports = class Checkout extends Page {
     get guestCheckoutHeader () { return $(GUEST_CHECKOUT_HEADER); }
 
     get guestCheckoutLoginButton () { return $(GUEST_CHECKOUT_LOGIN_BUTTON); }
+
+    get preOrderWarning () { return $(PRE_ORDER_WARNING); }
+
+    get checkoutErrorMessage () { return $(CHECKOUT_ERROR_MESSAGE); }
+
+    get errorMessageRetry () { return $(RETRY_BUTTON); }
+
+    get errorPageComponent () { return $(ERROR_PAGE_COMPONENT); }
+
+    get errorPageDescription () { return $(ERROR_PAGE_DESCRIPTION); }
+
+    get errorPageHeading () { return $(ERROR_PAGE_HEADING); }
+
+    get errorPageImage () { return $(ERROR_PAGE_IMAGE); }
 
     fields = {
         firstName: {
@@ -86,13 +107,19 @@ module.exports = class Checkout extends Page {
     open (checkout) {
         const serviceType = checkout.isValid ? `&knob-Service%20Type=${checkout.type}` : '&knob-Service%20Type=Invalid%20URL';
         const isLoggedIn = `&knob-Is%20User%20Logged%20In=${checkout.isAuthenticated}`;
+        const hasErrors = `&knob-Has%20Checkout%20Errors=${checkout.checkoutErrors}`;
+        const hasPreorderWarning = `&knob-Is%20ASAP%20available=${checkout.isPreOrderWarningDisplayed}`;
 
-        const url = `checkout-component${serviceType}${isLoggedIn}`;
+        const url = `checkout-component${serviceType}${isLoggedIn}${hasErrors}${hasPreorderWarning}`;
         super.openComponent('organism', url);
     }
 
     waitForComponent () {
         super.waitForComponent(this.component);
+    }
+
+    waitForErrorPageComponent () {
+        super.waitForComponent(this.errorPageComponent);
     }
 
     isComponentDisplayed () {
@@ -137,6 +164,34 @@ module.exports = class Checkout extends Page {
 
     isGuestCheckoutHeaderDisplayed () {
         return this.guestCheckoutHeader.isDisplayed();
+    }
+
+    isPreOrderWarningDisplayed () {
+        return this.preOrderWarning.isDisplayed();
+    }
+
+    isCheckoutErrorMessageDisplayed () {
+        return this.checkoutErrorMessage.isDisplayed();
+    }
+
+    clickRetryButton () {
+        return this.errorMessageRetry.click();
+    }
+
+    isErrorPageComponentDisplayed () {
+        return this.errorPageComponent.isDisplayed();
+    }
+
+    isErrorPageHeadingDisplayed () {
+        return this.errorPageHeading.isDisplayed();
+    }
+
+    isErrorPageDescriptionDisplayed () {
+        return this.errorPageDescription.isDisplayed();
+    }
+
+    isErrorPageImageDisplayed () {
+        return this.errorPageImage.isDisplayed();
     }